feat: update court and type attributes and create seeder for tennis court

+++ migrations/20231012194553-create-type.js
@@ -7,7 +7,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.SMALLINT
+        type: Sequelize.INTEGER
       },
       type: {
         type: Sequelize.STRING(50)
